requirements.txt
streamlit
openai
pinecone
python-dotenv
tiktoken
pypdf